feat(Termite.Terminal.PrimTTY): vendor erlang records
In certain environments (such as using hexpm/bob docker images) the
erlang source is removed from the release. This means it is not possible
to extract the record from an erlang source file.

To fix this, the state record used by prim_tty has been vendored into
the project, and the adapter uses the running otp version to determine
which record to load at compile time.